fix(github-graphql): guard agent rate-limit budget
