+++ .changeset/guard-agent-github-graphql-rate-limit.md
@@ -1,0 +1,5 @@
+---
+"@gh-symphony/cli": patch
+---
+
+Protect `github_graphql` agent calls with shared GitHub GraphQL rate-limit guards and record measured request cost for issue #481.
