+++ packages/tool-github-graphql/package.json
@@ -41,6 +41,10 @@
     "build": "tsc -p tsconfig.json",
     "lint": "eslint src --ext .ts",
     "test": "vitest run --passWithNoTests",
-    "typecheck": "tsc -p tsconfig.json --noEmit"
+    "typecheck": "tsc -p tsconfig.typecheck.json"
+  },
+  "dependencies": {
+    "@gh-symphony/core": "workspace:*",
+    "graphql": "^16.11.0"
   }
 }
