+++ packages/tool-github-graphql/src/index.ts
@@ -2,3 +2,4 @@ export * from "./tool.js";
 export * from "./mcp-server.js";
 export * from "./mcp-entry.js";
 export * from "./url-policy.js";
+export * from "./github-rate-limit.js";
