+++ packages/tool-github-graphql/src/tool.test.ts
@@ -6,6 +6,11 @@ import {
   DEFAULT_GITHUB_GRAPHQL_API_URL,
   createGitHubGraphQLMcpServerEntry,
 } from "./mcp-entry.js";
+import {
+  GitHubGraphQLRateLimitPolicy,
+  fingerprintGitHubToken,
+  githubGraphQLRateLimitPolicy,
+} from "./github-rate-limit.js";
 import { executeGitHubGraphQL, resolveGitHubGraphQLToken } from "./tool.js";
 
 const tempRoots: string[] = [];
@@ -295,6 +300,10 @@ describe("resolveGitHubGraphQLToken", () => {
 });
 
 describe("executeGitHubGraphQL", () => {
+  afterEach(() => {
+    githubGraphQLRateLimitPolicy.reset();
+  });
+
   it("posts to the public GitHub GraphQL API host", async () => {
     const fetchImpl = vi.fn().mockResolvedValue(
       new Response(JSON.stringify({ data: { viewer: { login: "octo" } } }), {
@@ -321,6 +330,137 @@ describe("executeGitHubGraphQL", () => {
         method: "POST",
       })
     );
+  });
+
+  it("instruments query cost and returns shared rate-limit metadata", async () => {
+    const fetchImpl = vi.fn().mockResolvedValue(
+      new Response(
+        JSON.stringify({
+          data: {
+            viewer: { login: "octo" },
+            rateLimit: {
+              cost: 7,
+              remaining: 4_992,
+              resetAt: "2026-08-03T01:00:00.000Z",
+            },
+          },
+        }),
+        {
+          status: 200,
+          headers: {
+            "x-ratelimit-limit": "5000",
+            "x-ratelimit-remaining": "4992",
+            "x-ratelimit-used": "8",
+            "x-ratelimit-reset": "1785718800",
+            "x-ratelimit-resource": "graphql",
+          },
+        }
+      )
+    );
+
+    const result = await executeGitHubGraphQL(
+      {
+        query: "query Viewer { viewer { login } }",
+      },
+      {
+        token: "ghs_static",
+      },
+      fetchImpl as typeof fetch
+    );
+
+    expect(result).toMatchObject({
+      rateLimits: {
+        source: "github",
+        cost: 7,
+        remaining: 4_992,
+        fieldRateLimits: {
+          cost: 7,
+          remaining: 4_992,
+        },
+        headerRateLimits: {
+          limit: 5_000,
+          used: 8,
+          resource: "graphql",
+        },
+      },
+    });
+
+    const request = fetchImpl.mock.calls[0]?.[1] as RequestInit;
+    const body = JSON.parse(String(request.body)) as { query: string };
+    expect(body.query).toContain("rateLimit");
+    expect(body.query).toContain("cost");
+    expect(body.query).toContain("remaining");
+    expect(body.query).toContain("resetAt");
+  });
+
+  it("blocks a request when the shared cached budget is exhausted", async () => {
+    const policy = new GitHubGraphQLRateLimitPolicy({
+      now: () => Date.parse("2026-08-03T00:00:00.000Z"),
+    });
+    policy.set(fingerprintGitHubToken("ghs_static"), {
+      source: "github",
+      remaining: 0,
+      resetAt: "2026-08-03T02:00:00.000Z",
+    });
+    const fetchImpl = vi.fn();
+
+    await expect(
+      executeGitHubGraphQL(
+        { query: "query Viewer { viewer { login } }" },
+        {
+          token: "ghs_static",
+          rateLimitPolicy: policy,
+        },
+        fetchImpl as typeof fetch
+      )
+    ).rejects.toMatchObject({
+      name: "GitHubGraphQLRateLimitError",
+    });
+    expect(fetchImpl).not.toHaveBeenCalled();
+  });
+
+  it("retries secondary rate limits with the shared bounded policy", async () => {
+    const sleep = vi.fn().mockResolvedValue(undefined);
+    const policy = new GitHubGraphQLRateLimitPolicy({
+      retryAttempts: 2,
+      retryBaseMs: 0,
+      sleep,
+    });
+    const fetchImpl = vi
+      .fn()
+      .mockResolvedValueOnce(
+        new Response(JSON.stringify({ message: "secondary rate limit" }), {
+          status: 429,
+          headers: { "retry-after": "0" },
+        })
+      )
+      .mockResolvedValueOnce(
+        new Response(
+          JSON.stringify({
+            data: {
+              viewer: { login: "octo" },
+              rateLimit: {
+                cost: 1,
+                remaining: 4_999,
+                resetAt: "2026-08-03T01:00:00.000Z",
+              },
+            },
+          }),
+          { status: 200 }
+        )
+      );
+
+    await expect(
+      executeGitHubGraphQL(
+        { query: "query Viewer { viewer { login } }" },
+        { token: "ghs_static", rateLimitPolicy: policy },
+        fetchImpl as typeof fetch
+      )
+    ).resolves.toMatchObject({
+      rateLimits: { cost: 1, remaining: 4_999 },
+    });
+    expect(fetchImpl).toHaveBeenCalledTimes(2);
+    expect(sleep).toHaveBeenCalledWith(0);
   });
 
   it("supports a configured public GHES GraphQL API URL", async () => {
