+++ packages/tool-github-graphql/tsconfig.typecheck.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true,
+    "rootDir": "..",
+    "baseUrl": ".",
+    "paths": {
+      "@gh-symphony/core": ["../core/src/index.ts"]
+    }
+  }
+}
