+++ packages/tool-github-graphql/vitest.config.ts
@@ -1,6 +1,15 @@
 import { defineConfig } from "vitest/config";
+import { dirname, resolve } from "node:path";
+import { fileURLToPath } from "node:url";
+
+const packageRoot = dirname(fileURLToPath(import.meta.url));
 
 export default defineConfig({
+  resolve: {
+    alias: {
+      "@gh-symphony/core": resolve(packageRoot, "../core/src/index.ts"),
+    },
+  },
   test: {
     environment: "node",
   },
