+++ packages/tracker-github/package.json
@@ -39,6 +39,7 @@
     "typecheck": "tsc -p tsconfig.typecheck.json"
   },
   "dependencies": {
-    "@gh-symphony/core": "workspace:*"
+    "@gh-symphony/core": "workspace:*",
+    "@gh-symphony/tool-github-graphql": "workspace:*"
   }
 }
